Rename spring.mvc to spring.webmvc to be consistent with name in Spring
