+++ agent/agent-plugins/spring-webmvc/pom.xml
@@ -7,7 +7,7 @@
     <version>1.0-SNAPSHOT</version>
     <relativePath>../../pom.xml</relativePath>
   </parent>
-  <artifactId>agent-plugin-spring-mvc</artifactId>
+  <artifactId>agent-plugin-spring-web</artifactId>
 
   <dependencies>
     <dependency>
@@ -18,12 +18,6 @@
       <groupId>org.springframework</groupId>
       <artifactId>spring-webmvc</artifactId>
       <version>4.3.12.RELEASE</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.github.openfeign</groupId>
-      <artifactId>feign-core</artifactId>
-      <version>10.8</version>
       <scope>provided</scope>
     </dependency>
   </dependencies>
