+++ agent/agent-plugins/spring-webmvc/src/main/java/org/bithon/agent/plugin/spring/webmvc/rs/RestTemplate$HandleResponse.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.agent.plugin.spring.mvc;
+package org.bithon.agent.plugin.spring.webmvc.rs;
 
 import org.bithon.agent.instrumentation.aop.context.AopContext;
 import org.bithon.agent.instrumentation.aop.interceptor.declaration.BeforeInterceptor;
